+++ CHANGELOG.md
@@ -5,3 +5,4 @@ All notable changes to this project will be documented in this file.
 ## 0.4.0
 
 - [Feature] Add missing aggregation methods for Series class: `argMax()`, `argMin()`, `nanMax()`, `nanMin()`, `maxBy()`, `minBy()`, `mode()`, `implode()`
+- Add `Series::getFlags()` method to retrieve sorting and optimization flags